refactored code for maintainability and readability
