+++ src/lib/panes.js
